requirements.txt
# Minimal requirements for voice AI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
twilio==8.10.0
requests==2.31.0